(feat): Add module to validate objectids for all routes that work with req.params.id - this is a custom module named validateobjectid.js

+++ routes/bookingRouter.js
@@ -3,7 +3,7 @@ const router = express.Router();
 const booking = require('../models/booking.model');
 const chalk = require('chalk');
 const mongoose = require('mongoose');
-
+const validateId = require('../validateObjectId');
 
 router.get('/', function (req, res) {
   booking.find({}, (error, bookings) => {
@@ -15,6 +15,9 @@ router.get('/', function (req, res) {
 });
 
 router.get('/:id', function (req, res) {
+  if (!validateId(req.params.id)) {
+    return res.status(400).json('Invalid booking id');
+  }
   booking.findById(req.params.id, (error, booking) => {
     if (error) {
       return res.status(500).json(error);
@@ -39,6 +42,9 @@ router.post('/', function (req, res) {
 });
 
 router.delete('/:id', function (req, res) {
+  if (!validateId(req.params.id)) {
+    return res.status(400).json('Invalid booking id');
+  }
   booking.findByIdAndRemove(req.params.id, (error, booking) => {
     if (error) {
       return res.status(500).json(error);
@@ -48,10 +54,16 @@ router.delete('/:id', function (req, res) {
 });
 
 router.patch('/:id', function (req, res) {
+  if (!validateId(req.params.id)) {
+    return res.status(400).json('Invalid booking id');
+  }
   return res.status(200).send('booking updated');
 });
 
-router.put('/:id', function(req, res) {
+router.put('/:id', function (req, res) {
+  if (!validateId(req.params.id)) {
+    return res.status(400).json('Invalid booking id');
+  }
   return res.status(200).send('booking updated');
 });
