+++ routes/equipmentRouter.js
@@ -3,6 +3,7 @@ const router = express.Router();
 const equipment = require('../models/equipment.model');
 const chalk = require('chalk');
 const mongoose = require('mongoose');
+const validateId = require('../validateObjectId');
 
 router.get('/', function (req, res) {
   equipment.find({}, (error, equipments) => {
@@ -14,6 +15,9 @@ router.get('/', function (req, res) {
 });
 
 router.get('/:id', function (req, res) {
+  if (!validateId(req.params.id)) {
+    return res.status(400).json('Invalid equipment id');
+  }
   equipment.findById(req.params.id, (error, equipment) => {
     if (error) {
       return res.status(500).json(error);
@@ -35,6 +39,9 @@ router.post('/', function (req, res) {
 });
 
 router.delete('/:id', function (req, res) {
+  if (!validateId(req.params.id)) {
+    return res.status(400).json('Invalid equipment id');
+  }
   equipment.findByIdAndRemove(req.params.id, (error, equipment) => {
     if (error) {
       return res.status(500).json(error);
@@ -44,10 +51,16 @@ router.delete('/:id', function (req, res) {
 });
 
 router.patch('/:id', function (req, res) {
+  if (!validateId(req.params.id)) {
+    return res.status(400).json('Invalid equipment id');
+  }
   return res.status(200).send('equipment updated');
 });
 
 router.put('/:id', function (req, res) {
+  if (!validateId(req.params.id)) {
+    return res.status(400).json('Invalid equipment id');
+  }
   return res.status(200).send('equipment updated');
 });
